Turn off dark mode
Added the fix_info_plist.py script to turn off dark mode. This is required when the build is done on an older Mac OS (e.g. El Capitan) and the app is run on Mojave or later. With the current GUI colors, dark mode no good. This could be fixed in the app, but I'm currently trying not to touch that.